@@ -1,7 +1,5 @@
 package com.liaoyb.qingqing.goods.client;
 
-import java.io.IOException;
-
 import org.springframework.cloud.security.oauth2.client.feign.OAuth2FeignRequestInterceptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.oauth2.client.DefaultOAuth2ClientContext;
@@ -18,7 +16,7 @@ public class OAuth2InterceptedFeignConfiguration {
     }
 
     @Bean(name = "oauth2RequestInterceptor")
-    public RequestInterceptor getOAuth2RequestInterceptor() throws IOException {
+    public RequestInterceptor getOAuth2RequestInterceptor() {
         return new OAuth2FeignRequestInterceptor(new DefaultOAuth2ClientContext(), loadBalancedResourceDetails);
     }
 }
