@@ -5,13 +5,8 @@ package com.liaoyb.qingqing.goods.config;
  */
 public final class Constants {
 
-    // Regex for acceptable logins
-    public static final String LOGIN_REGEX = "^[_.@A-Za-z0-9-]*$";
-
     public static final String SYSTEM_ACCOUNT = "system";
-    public static final String ANONYMOUS_USER = "anonymoususer";
-    public static final String DEFAULT_LANGUAGE = "zh-cn";
-    
+
     private Constants() {
     }
 }
