@@ -1,18 +1,22 @@
 package com.liaoyb.qingqing.goods.config;
 
 import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.cloud.openfeign.FeignClientsConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @Configuration
 @EnableFeignClients(basePackages = "com.liaoyb.qingqing.goods")
+@Import(FeignClientsConfiguration.class)
 public class FeignConfiguration {
 
     /**
-     * Set the Feign specific log level to log client REST requests
+     * Set the Feign specific log level to log client REST requests.
      */
     @Bean
     feign.Logger.Level feignLoggerLevel() {
         return feign.Logger.Level.BASIC;
     }
+
 }
