@@ -34,9 +34,9 @@ public class OAuth2Properties {
         /**
          * Defines the cookie domain. If specified, cookies will be set on this domain.
          * If not configured, then cookies will be set on the top-level domain of the
-         * request you sent, i.e. if you send a request to <code>app1.your-domain.com</code>,
-         * then cookies will be set <code>on .your-domain.com</code>, such that they
-         * are also valid for <code>app2.your-domain.com</code>.
+         * request you sent, i.e. if you send a request to {@code app1.your-domain.com},
+         * then cookies will be set on {@code .your-domain.com}, such that they
+         * are also valid for {@code app2.your-domain.com}.
          */
         private String cookieDomain;
 
