@@ -1,7 +1,8 @@
 package com.liaoyb.qingqing.goods.web.rest.errors;
 
-import com.liaoyb.qingqing.goods.web.rest.util.HeaderUtil;
+import io.github.jhipster.web.util.HeaderUtil;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.ConcurrencyFailureException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -25,7 +26,7 @@ import java.util.stream.Collectors;
 
 /**
  * Controller advice to translate the server side exceptions to client-friendly json structures.
- * The error response follows RFC7807 - Problem Details for HTTP APIs (https://tools.ietf.org/html/rfc7807)
+ * The error response follows RFC7807 - Problem Details for HTTP APIs (https://tools.ietf.org/html/rfc7807).
  */
 @ControllerAdvice
 public class ExceptionTranslator implements ProblemHandling {
@@ -35,8 +36,11 @@ public class ExceptionTranslator implements ProblemHandling {
     private static final String PATH_KEY = "path";
     private static final String VIOLATIONS_KEY = "violations";
 
+    @Value("${jhipster.clientApp.name}")
+    private String applicationName;
+
     /**
-     * Post-process the Problem payload to add the message key for the front-end if needed
+     * Post-process the Problem payload to add the message key for the front-end if needed.
      */
     @Override
     public ResponseEntity<Problem> process(@Nullable ResponseEntity<Problem> entity, NativeWebRequest request) {
@@ -98,7 +102,7 @@ public class ExceptionTranslator implements ProblemHandling {
 
     @ExceptionHandler
     public ResponseEntity<Problem> handleBadRequestAlertException(BadRequestAlertException ex, NativeWebRequest request) {
-        return create(ex, request, HeaderUtil.createFailureAlert(ex.getEntityName(), ex.getErrorKey(), ex.getMessage()));
+        return create(ex, request, HeaderUtil.createFailureAlert(applicationName, true, ex.getEntityName(), ex.getErrorKey(), ex.getMessage()));
     }
 
     @ExceptionHandler
